worker/requirements.txt
redis
requests
python-dotenv
# Add these for the real model:
# torch
# torchaudio
# audiocraft
# boto3
